requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-Session==0.8.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Marshmallow==3.20.2
qrcode[pil]==7.4.2
Pillow==10.1.0
phonenumbers==8.13.26
reportlab==4.0.7
weasyprint==60.2
openpyxl==3.1.2
pandas==2.1.4
matplotlib==3.8.2
plotly==5.17.0
redis==5.0.1
python-dotenv==1.0.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0
bleach==6.1.0
